refactor: http.exceptions 패키지 추출
