+++ src/main/java/http/exceptions/FailToConvertMessageException.java
@@ -1,4 +1,4 @@
-package http.supoort;
+package http.exceptions;
 
 public class FailToConvertMessageException extends RuntimeException {
     public FailToConvertMessageException() {
