+++ src/main/java/http/model/HttpMethod.java
@@ -1,6 +1,6 @@
 package http.model;
 
-import http.supoort.IllegalHttpRequestException;
+import http.exceptions.IllegalHttpRequestException;
 
 public enum HttpMethod {
     GET,
