+++ src/main/java/http/model/HttpUri.java
@@ -1,6 +1,6 @@
 package http.model;
 
-import http.supoort.IllegalHttpRequestException;
+import http.exceptions.IllegalHttpRequestException;
 
 import java.util.Objects;
