+++ src/main/java/utils/FileIoUtils.java
@@ -1,5 +1,6 @@
 package utils;
 
+import http.exceptions.NoSuchResource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
