+++ src/test/java/http/model/HttpProtocolsTest.java
@@ -1,6 +1,6 @@
 package http.model;
 
-import http.supoort.IllegalHttpRequestException;
+import http.exceptions.IllegalHttpRequestException;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Java6Assertions.assertThatThrownBy;
