+++ src/test/java/http/supoort/HttpRequestParserTest.java
@@ -1,5 +1,6 @@
 package http.supoort;
 
+import http.exceptions.IllegalHttpRequestException;
 import http.model.*;
 import org.junit.jupiter.api.Test;
